backend/requirements.txt
# FastAPI and server dependencies
uvicorn[standard]==0.24.0
fastapi==0.104.1
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database - MySQL with SQLAlchemy (and SQLite for local development)
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
mysqlclient==2.2.4

# File storage and cloud services
boto3==1.34.162
botocore==1.34.162

# Document processing
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

# NLP and text processing
spacy==3.7.6

# HTTP requests and web scraping
requests==2.31.0
beautifulsoup4==4.12.3

# Rate limiting
slowapi==0.1.9
redis==5.0.1

# Email notifications
sendgrid==6.11.0

# Environment and configuration
python-dotenv==1.0.1

# Logging and monitoring
structlog==23.2.0

# AI/ML
openai==1.12.0

# Development dependencies
ipykernel==6.29.5